feat(components): decompose checkbox file into three files
